+++ src/pyzet/main.py
@@ -179,7 +179,6 @@ def _get_parser() -> tuple[ArgumentParser, dict[str, ArgumentParser]]:
         default=C.DEFAULT_REMOTE_NAME,
         help='name of git repo remote (default: %(default)s)',
     )
-    _add_git_cmd_options(remote_parser, 'remote')
 
     define_sample_config_cli(subparsers)
 
@@ -308,7 +307,7 @@ def process_yaml(
 
 
 def get_remote_url(args: Namespace, config: Config) -> int:
-    print(get_git_remote_url(config, args.origin, args.options))
+    print(get_git_remote_url(config, args.origin))
     return 0