+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.10.0] - 2024-01-17
+
 ### Added
 
 * Ask to create a new zettel if none are found by grep in `pyzet edit` & `pyzet show`.
@@ -142,7 +144,8 @@ Initial release.
 
 <!-- Links -->
 
-[Unreleased]: https://github.com/tpwo/pyzet/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/tpwo/pyzet/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/tpwo/pyzet/releases/tag/v0.10.0
 [0.9.0]: https://github.com/tpwo/pyzet/releases/tag/v0.9.0
 [0.8.0]: https://github.com/tpwo/pyzet/releases/tag/v0.8.0
 [0.7.0]: https://github.com/tpwo/pyzet/releases/tag/v0.7.0